L1 hashing psw/backend/google-cloud-functions/mfa/requirements.txt
flask==2.3.3
google-cloud-firestore==2.13.1
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
